requirements.txt
Flask
numpy
matplotlib
Pillow
